+++ apis/auth/apis.ts
@@ -1,5 +1,4 @@
 import publicApi from '../config/publicApi';
-import type { AxiosResponse } from 'axios';
 
 import type {
   EmailRequest,
@@ -10,6 +9,7 @@ import type {
   SignUpRequest,
   SignUpResponse,
 } from './type';
+import type { AxiosResponse } from 'axios';
 
 const TEST_ID = {
   email: 'testy54@soongsil.ac.kr',
