+++ apis/auth/index.ts
@@ -2,3 +2,4 @@ export * from './apis';
 export * from './keys';
 export * from './mutations';
 export * from './queries';
+export * from './type';
