+++ apis/common/type.ts
@@ -1,0 +1,1 @@
+export type ExampleType = {};
