+++ apis/groups/apis.ts
@@ -1,4 +1,4 @@
-import { GroupResponse } from './type';
+import { GroupResponse } from '.';
 import privateApi from '../config/privateApi';
 
 export const getGroups = (pageNum: number) => {
